# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MMA Tournament API
  version: 1.0.0
  description: |
    API for managing MMA tournament data including athletes, events, rankings, and matches.
    Supports multiple tournaments (UFC, Lion Championship, etc.) with comprehensive CRUD operations.
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: http://localhost:4000/api
    description: Local development API base
...
paths:
...
  /api/rankings:
    get:
      summary: Get all rankings
      description: Retrieve P4P rankings and all division rankings for a tournament
      tags:
        - Rankings
      parameters:
        - name: tournament
          in: query
          description: Tournament identifier
          schema:
            type: string
            default: ufc
            enum: [ufc, lion]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    type: string
                    example: ufc
                  p4pRankings:
                    type: array
                    items:
                      $ref: '#/components/schemas/P4PRanking'
                  divisions:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        champion:
                          $ref: '#/components/schemas/Ranking'
                        rankings:
                          type: array
                          items:
                            $ref: '#/components/schemas/Ranking'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new ranking
      description: Add a new division ranking
      tags:
        - Rankings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RankingInput'
      responses:
        '201':
          description: Ranking created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: Ranking created successfully
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rankings/p4p:
    get:
      summary: Get P4P rankings
      description: Retrieve pound-for-pound rankings for a tournament
      tags:
        - Rankings
      parameters:
        - name: tournament
          in: query
          description: Tournament identifier
          schema:
            type: string
            default: ufc
            enum: [ufc, lion]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    type: string
                    example: ufc
                  p4pRankings:
                    type: array
                    items:
                      $ref: '#/components/schemas/P4PRanking'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new P4P ranking
      description: Add a new pound-for-pound ranking
      tags:
        - Rankings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/P4PRankingInput'
      responses:
        '201':
          description: P4P ranking created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: P4P ranking created successfully
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rankings/{division}:
    get:
      summary: Get rankings for specific division
      description: Retrieve champion and rankings for a specific weight division
      tags:
        - Rankings
      parameters:
        - name: division
          in: path
          required: true
          description: Weight division name
          schema:
            type: string
            example: lightweight
        - name: tournament
          in: query
          description: Tournament identifier
          schema:
            type: string
            default: ufc
            enum: [ufc, lion]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournament:
                    type: string
                    example: ufc
                  division:
                    type: string
                    example: lightweight
                  champion:
                    $ref: '#/components/schemas/Ranking'
                  rankings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ranking'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/promotions:
    get:
      summary: Get all promotions
      description: Retrieve all promotions
      tags:
        - Promotions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 2
                  promotions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Promotion'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create new promotion
      description: Add a new promotion to the database
      tags:
        - Promotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '201':
          description: Promotion created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/promotions/{id}:
    get:
      summary: Get promotion by ID
      description: Retrieve a specific promotion by its ID
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update promotion
      description: Update an existing promotion's information
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionInput'
      responses:
        '200':
          description: Promotion updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete promotion
      description: Remove a promotion from the database
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
      responses:
        '200':
          description: Promotion deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Promotion deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/promotions/{id}/events:
    post:
      summary: Add event to promotion
      description: Add an event ID to a promotion's events array
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
              properties:
                event_id:
                  type: string
                  example: ufc-310
      responses:
        '201':
          description: Event added to promotion successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event added to promotion successfully
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          description: Promotion not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Promotion not found
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Remove event from promotion
      description: Remove an event ID from a promotion's events array
      tags:
        - Promotions
      parameters:
        - $ref: '#/components/parameters/PromotionId'
        - name: event_id
          in: query
          required: true
          description: Event ID to remove
          schema:
            type: string
            example: ufc-310
      responses:
        '200':
          description: Event removed from promotion successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Event removed from promotion successfully
                  promotion:
                    $ref: '#/components/schemas/Promotion'
        '404':
          description: Promotion or event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Promotion not found
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    Ranking:
      type: object
      required:
        - tournament
        - division
        - rank
        - athlete_name
      properties:
        id:
          type: integer
          example: 1
        tournament:
          type: string
          example: ufc
        division:
          type: string
          example: lightweight
        rank:
          type: integer
          example: 1
        athlete_id:
          type: integer
          nullable: true
          example: 1
        athlete_name:
          type: string
          example: Test Fighter
        is_champion:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
        updated_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
    RankingInput:
      type: object
      required:
        - tournament
        - division
        - rank
        - athlete_name
      properties:
        tournament:
          type: string
          example: ufc
        division:
          type: string
          example: lightweight
        rank:
          type: integer
          example: 1
        athlete_id:
          type: integer
          example: 1
        athlete_name:
          type: string
          example: Test Fighter
        is_champion:
          type: boolean
          example: false
    P4PRanking:
      type: object
      required:
        - tournament
        - rank
        - athlete_name
      properties:
        id:
          type: integer
          example: 1
        tournament:
          type: string
          example: ufc
        rank:
          type: integer
          example: 1
        athlete_id:
          type: integer
          nullable: true
          example: 1
        athlete_name:
          type: string
          example: Test Fighter
        created_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
        updated_at:
          type: string
          format: date-time
          example: '2025-11-24 01:15:09'
    P4PRankingInput:
      type: object
      required:
        - tournament
        - rank
        - athlete_name
      properties:
        tournament:
          type: string
          example: ufc
        rank:
          type: integer
          example: 1
        athlete_id:
          type: integer
          example: 1
        athlete_name:
          type: string
          example: Test Fighter
    Promotion:
      type: object
      required:
        - id
        - name
        - subtitle
        - theme
        - color
        - events
      properties:
        id:
          type: string
          example: ufc
        name:
          type: string
          example: UFC
        subtitle:
          type: string
          example: Ultimate Fighting Championship
        theme:
          type: string
          example: ufc-theme
        color:
          type: string
          example: '#d20a0a'
        events:
          type: array
          items:
            type: string
          example: ['ufc-309', 'ufc-308']
    PromotionInput:
      type: object
      required:
        - id
        - name
        - subtitle
        - theme
        - color
      properties:
        id:
          type: string
          example: ufc
        name:
          type: string
          example: UFC
        subtitle:
          type: string
          example: Ultimate Fighting Championship
        theme:
          type: string
          example: ufc-theme
        color:
          type: string
          example: '#d20a0a'
        events:
          type: array
          items:
            type: string
          example: ['ufc-309', 'ufc-308']
  parameters:
...
    PromotionId:
      name: id
      in: path
      required: true
      description: Promotion ID
      schema:
        type: string
        example: ufc
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Athlete not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Failed to fetch athletes